+++ client/http-client/src/rpc_service.rs
@@ -5,8 +5,8 @@ use hyper::body::Bytes;
 use jsonrpsee_core::{
 	BoxError, JsonRawValue,
 	client::Error,
+	method_response::{BatchResponseBuilder, MethodResponse, ResponsePayload},
 	middleware::{Notification, RpcServiceT},
-	server::{BatchResponseBuilder, MethodResponse, ResponsePayload},
 };
 use jsonrpsee_types::{Id, Response, ResponseSuccess};
 use tower::Service;
