+++ core/src/server/helpers.rs
@@ -24,58 +24,12 @@
 // IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::io;
 use std::time::Duration;
 
 use jsonrpsee_types::{ErrorCode, ErrorObject, Id, InvalidRequest, Response, ResponsePayload};
 use tokio::sync::mpsc;
 
 use super::{DisconnectError, SendTimeoutError, SubscriptionMessage, TrySendError};
-
-/// Bounded writer that allows writing at most `max_len` bytes.
-///
-/// ```
-///    use std::io::Write;
-///
-///    use jsonrpsee_core::server::helpers::BoundedWriter;
-///
-///    let mut writer = BoundedWriter::new(10);
-///    (&mut writer).write("hello".as_bytes()).unwrap();
-///    assert_eq!(std::str::from_utf8(&writer.into_bytes()).unwrap(), "hello");
-/// ```
-#[derive(Debug, Clone)]
-pub struct BoundedWriter {
-	max_len: usize,
-	buf: Vec<u8>,
-}
-
-impl BoundedWriter {
-	/// Create a new bounded writer.
-	pub fn new(max_len: usize) -> Self {
-		Self { max_len, buf: Vec::with_capacity(128) }
-	}
-
-	/// Consume the writer and extract the written bytes.
-	pub fn into_bytes(self) -> Vec<u8> {
-		self.buf
-	}
-}
-
-impl io::Write for &mut BoundedWriter {
-	fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-		let len = self.buf.len() + buf.len();
-		if self.max_len >= len {
-			self.buf.extend_from_slice(buf);
-			Ok(buf.len())
-		} else {
-			Err(io::Error::new(io::ErrorKind::OutOfMemory, "Memory capacity exceeded"))
-		}
-	}
-
-	fn flush(&mut self) -> io::Result<()> {
-		Ok(())
-	}
-}
 
 /// Sink that is used to send back the result to the server for a specific method.
 #[derive(Clone, Debug)]
@@ -169,28 +123,5 @@ pub fn prepare_error(data: &[u8]) -> (Id<'_>, ErrorCode) {
 	match serde_json::from_slice::<InvalidRequest>(data) {
 		Ok(InvalidRequest { id }) => (id, ErrorCode::InvalidRequest),
 		Err(_) => (Id::Null, ErrorCode::ParseError),
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use crate::server::BoundedWriter;
-	use jsonrpsee_types::{Id, Response, ResponsePayload};
-
-	#[test]
-	fn bounded_serializer_work() {
-		let mut writer = BoundedWriter::new(100);
-		let result = ResponsePayload::success(&"success");
-		let rp = &Response::new(result, Id::Number(1));
-
-		assert!(serde_json::to_writer(&mut writer, rp).is_ok());
-		assert_eq!(String::from_utf8(writer.into_bytes()).unwrap(), r#"{"jsonrpc":"2.0","id":1,"result":"success"}"#);
-	}
-
-	#[test]
-	fn bounded_serializer_cap_works() {
-		let mut writer = BoundedWriter::new(100);
-		// NOTE: `"` is part of the serialization so 101 characters.
-		assert!(serde_json::to_writer(&mut writer, &"x".repeat(99)).is_err());
 	}
 }
